requirements.txt
configobj
ffmpy
mutagen
pillow
pydub
